+++ CHANGELOG.md
@@ -4,17 +4,20 @@ change log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `on` event matcher expecting a predicate of the event.
+- `on-every` event matcher, matching all events.
 
 ## [0.6.2] — 2019-08-22
 ### Added
-- Codox
+- Codox.
 
 ## [0.6.1] — 2019-08-22
 ### Changed
-- Renamed `on` to `on-type` (breaking)
+- Renamed `on` to `on-type` (breaking).
 
 ## [0.6.0] - 2019-05-22
-- Initial release
+- Initial release.
 
 [0.6.0]: https://github.com/your-name/vent/compare/0.1.0...0.6.0
 [0.6.1]: https://github.com/your-name/vent/compare/0.6.0...0.6.1